refactor: extract an amqp package
