+++ datashare-app/src/main/java/org/icij/datashare/com/bus/amqp/AmqpChannel.java
@@ -1,4 +1,4 @@
-package org.icij.datashare.com.bus;
+package org.icij.datashare.com.bus.amqp;
 
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.ConfirmCallback;
