+++ datashare-app/src/main/java/org/icij/datashare/com/bus/amqp/Deserializer.java
@@ -1,4 +1,4 @@
-package org.icij.datashare.com.bus;
+package org.icij.datashare.com.bus.amqp;
 
 import java.io.IOException;
